Update camserver.py
Camera stream currently works although needs to be rotated 270 degrees and swapped to the noir tuning file for the imx219 sensor. 

Starting a recording now works as well as stopping a recording saved the video under /home/cm4/cam/static/video/ as a h264 file. Working on having them saved as a mkv or mp4. 

Snapping a picture now works and saves a picture to /home/cm4/cam/static/picture/

The files button now takes you to a page where you can live view the picture files taken by the snap feature where currently one or less thumbnails may be visible. 

Have not updated files.html to allow for downloading and expanding view of items on the pi.