Create files.html
Added a live file viewer for saved photos and videos. 

Thumbnails currently do not work but it is being worked on.
Files are currently not clickable.
Files are currently not able to be downloaded.

Issue with camserver.py saving the files with a hour and second time format inserting colons : and windows does not like colons in file names.